+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `AWS_DEFAULT_REGION="localhost"` var. in .env.test for omniqueue queue testing.
 - Added basic rust-toolchain support.
 - Tests for DA job.
+- Tests for prover client.
 
 ## Changed
 
@@ -30,3 +31,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `fetch_from_test` argument
 
 ## Fixed
+
+- Get Fact Info logic.